Remove all server side code
Fixes #10. Some code remains in [./src/codec.rs] in case of breaking the module out into a subcrate later.
